requirements.txt
absl-py==2.1.0
accelerate==0.26.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
av==14.2.0
bert-score==0.3.13
bitsandbytes==0.43.0
braceexpand==0.1.7
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.3.2
debugpy==1.8.13
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
dill==0.3.8
docker-pycreds==0.4.0
et_xmlfile==2.0.0
evaluate==0.4.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.29.3
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.1
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.26.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.2.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
portalocker==3.1.1
progressbar2==4.5.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==3.20.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycocotools==2.0.8
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
qwen-vl-utils==0.0.10
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
responses==0.18.0
rich==13.9.4
rouge_score==0.1.2
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.1.99
sentry-sdk==2.24.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
supervision==0.25.1
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.13
tokenizers==0.15.2
tomli==2.2.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.37.2
triton==2.3.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
visual-genome==1.1.1
wandb==0.19.8
wcwidth==0.2.13
webdataset==0.2.48
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
